refactor(settings): colocate settings components
