+++ frontend/src/pages/settings/email-notice/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { gettext } from '../../utils/constants';
-import { seafileAPI } from '../../utils/seafile-api';
-import { Utils } from '../../utils/utils';
-import toaster from '../toast';
+import toaster from '@/components/toast';
+import { gettext } from '../../../utils/constants';
+import { seafileAPI } from '../../../utils/seafile-api';
+import { Utils } from '../../../utils/utils';
 
 const {
   fileUpdatesEmailInterval,
