+++ frontend/src/pages/settings/index.js
@@ -1,0 +1,211 @@
+import React from 'react';
+import MediaQuery from 'react-responsive';
+import { Modal } from 'reactstrap';
+import UserSetPassword from '../../components/dialog/user-password-widgets/user-set-password-dialog';
+import UserUpdatePassword from '../../components/dialog/user-password-widgets/user-update-password-dialog';
+import toaster from '../../components/toast';
+import { isPro, isDBSqlite3, gettext } from '../../utils/constants';
+import { seafileAPI } from '../../utils/seafile-api';
+import { Utils } from '../../utils/utils';
+import DeleteAccount from './delete-account';
+import EmailNotice from './email-notice';
+import LanguageSetting from './language-setting';
+import LinkedDevices from './linked-devices';
+import ListInAddressBook from './list-in-address-book';
+import MainPanel from './main-panel';
+import SidePanel from './side-panel';
+import SocialLogin from './social-login';
+import SocialLoginDingtalk from './social-login-dingtalk';
+import SocialLoginSAML from './social-login-saml';
+import SocialLoginWeixin from './social-login-weixin';
+import TopToolbar from './top-toolbar';
+import TwoFactorAuthentication from './two-factor-auth';
+import UserAvatarForm from './user-avatar-form';
+import UserBasicInfoForm from './user-basic-info-form';
+import WebAPIAuthToken from './web-api-auth-token';
+import WebdavPassword from './webdav-password';
+
+import '../../css/layout.css';
+import '../../css/toolbar.css';
+import '../../css/search.css';
+import '../../css/user-settings.css';
+
+const {
+  canUpdatePassword, passwordOperationText,
+  enableGetAuthToken,
+  enableWebdavSecret,
+  enableAddressBook,
+  twoFactorAuthEnabled,
+  enableWechatWork,
+  enableDingtalk,
+  enableWeixin,
+  isOrgContext,
+  enableADFS,
+  enableMultiADFS,
+  enableDeleteAccount,
+  userUnusablePassword
+} = window.app.pageOptions;
+
+class Settings extends React.Component {
+
+  constructor(props) {
+    super(props);
+    this.sideNavItems = [
+      { show: true, href: '#user-basic-info', text: gettext('Profile') },
+      { show: canUpdatePassword, href: '#update-user-passwd', text: gettext('Password') },
+      { show: enableGetAuthToken, href: '#get-auth-token', text: gettext('Web API Auth Token') },
+      { show: enableWebdavSecret, href: '#update-webdav-passwd', text: gettext('WebDAV Access') },
+      { show: enableAddressBook, href: '#list-in-address-book', text: gettext('Global Address Book') },
+      { show: true, href: '#lang-setting', text: gettext('Language') },
+      { show: isPro, href: '#email-notice', text: gettext('Email Notification') },
+      { show: twoFactorAuthEnabled, href: '#two-factor-auth', text: gettext('Two-Factor Authentication') },
+      { show: (enableWechatWork || enableDingtalk || enableWeixin || enableADFS || (enableMultiADFS && isOrgContext)), href: '#social-auth', text: gettext('Single Sign On (SSO)') },
+      { show: true, href: '#linked-devices', text: gettext('Linked Devices') },
+      { show: enableDeleteAccount, href: '#del-account', text: gettext('Delete Account') },
+    ];
+
+    this.state = {
+      isSidePanelClosed: false,
+      curItemID: this.sideNavItems[0].href.substr(1),
+      isSetPasswordDialogOpen: false,
+      isUpdatePasswordDialogOpen: false
+
+    };
+  }
+
+  componentDidMount() {
+    seafileAPI.getUserInfo().then((res) => {
+      this.setState({
+        userInfo: res.data
+      });
+    }).catch(error => {
+      let errMessage = Utils.getErrorMsg(error);
+      toaster.danger(errMessage);
+    });
+  }
+
+  updateUserInfo = (data) => {
+    seafileAPI.updateUserInfo(data).then((res) => {
+      this.setState({
+        userInfo: res.data
+      });
+      toaster.success(gettext('User info updated'));
+    }).catch((error) => {
+      let errMessage = Utils.getErrorMsg(error);
+      toaster.danger(errMessage);
+    });
+  };
+
+  handleContentScroll = (e) => {
+    const scrollTop = e.target.scrollTop;
+    const scrolled = this.sideNavItems.filter((item, index) => {
+      return item.show && document.getElementById(item.href.substr(1)).offsetTop - 45 < scrollTop;
+    });
+    if (scrolled.length) {
+      this.setState({
+        curItemID: scrolled[scrolled.length - 1].href.substr(1)
+      });
+    }
+  };
+
+  onCloseSidePanel = () => {
+    this.setState({
+      isSidePanelClosed: !this.state.isSidePanelClosed
+    });
+  };
+
+  onShowSidePanel = () => {
+    this.setState({
+      isSidePanelClosed: !this.state.isSidePanelClosed
+    });
+  };
+
+  toggleSidePanel = () => {
+    this.setState({
+      isSidePanelClosed: !this.state.isSidePanelClosed
+    });
+  };
+
+  toggleSetPassword = () => {
+    this.setState({ isSetPasswordDialogOpen: !this.state.isSetPasswordDialogOpen });
+  };
+
+  toggleUpdatePassword = () => {
+    this.setState({ isUpdatePasswordDialogOpen: !this.state.isUpdatePasswordDialogOpen });
+  };
+
+  togglePassword = () => {
+    if (userUnusablePassword) {
+      this.toggleSetPassword();
+    } else {
+      this.toggleUpdatePassword();
+    }
+  };
+
+  render() {
+    const { isSidePanelClosed } = this.state;
+    return (
+      <div id="main" className="h-100">
+        <SidePanel
+          isSidePanelClosed={isSidePanelClosed}
+          onCloseSidePanel={this.onCloseSidePanel}
+          data={this.sideNavItems}
+          curItemID={this.state.curItemID}
+        />
+        <MainPanel>
+          <TopToolbar onShowSidePanel={this.onShowSidePanel} showSearch={false} />
+          <div className="main-panel-center flex-row">
+            <div className="cur-view-container">
+              <div className="cur-view-path">
+                <h3 className="sf-heading m-0">{gettext('Settings')}</h3>
+              </div>
+              <div className="cur-view-content content position-relative" onScroll={this.handleContentScroll}>
+                <div id="user-basic-info" className="setting-item">
+                  <h3 className="setting-item-heading">{gettext('Profile Setting')}</h3>
+                  <UserAvatarForm />
+                  {this.state.userInfo && <UserBasicInfoForm userInfo={this.state.userInfo} updateUserInfo={this.updateUserInfo} />}
+                </div>
+                {canUpdatePassword &&
+                <div id="update-user-passwd" className="setting-item">
+                  <h3 className="setting-item-heading">{gettext('Password')}</h3>
+                  <button className="btn btn-outline-primary mb-2" onClick={this.togglePassword}>{passwordOperationText}</button>
+                </div>
+                }
+                {enableGetAuthToken && <WebAPIAuthToken />}
+                {enableWebdavSecret && <WebdavPassword />}
+                {enableAddressBook && this.state.userInfo &&
+                <ListInAddressBook userInfo={this.state.userInfo} updateUserInfo={this.updateUserInfo} />}
+                <LanguageSetting />
+                {(isPro || !isDBSqlite3) && <EmailNotice />}
+                {twoFactorAuthEnabled && <TwoFactorAuthentication />}
+                {enableWechatWork && <SocialLogin />}
+                {enableDingtalk && <SocialLoginDingtalk />}
+                {enableWeixin && <SocialLoginWeixin />}
+                {(enableADFS || (enableMultiADFS && isOrgContext)) && <SocialLoginSAML />}
+                <LinkedDevices />
+                {enableDeleteAccount && <DeleteAccount />}
+              </div>
+            </div>
+          </div>
+        </MainPanel>
+        <MediaQuery query="(max-width: 767.8px)">
+          <Modal zIndex="1030" isOpen={!isSidePanelClosed} toggle={this.toggleSidePanel} contentClassName="d-none"></Modal>
+        </MediaQuery>
+        {this.state.isSetPasswordDialogOpen && (
+          <UserSetPassword
+            toggle={this.toggleSetPassword}
+          />
+        )}
+        {this.state.isUpdatePasswordDialogOpen && (
+          <UserUpdatePassword
+            toggle={this.toggleUpdatePassword}
+          />
+        )}
+      </div>
+
+
+    );
+  }
+}
+
+export default Settings;
