+++ frontend/src/pages/settings/linked-devices/index.js
@@ -4,13 +4,13 @@ import classnames from 'classnames';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
 import PropTypes from 'prop-types';
-import ConfirmUnlinkDeviceDialog from '../../components/dialog/confirm-unlink-device';
-import MobileItemMenu from '../../components/mobile-item-menu';
-import toaster from '../../components/toast';
-import { gettext } from '../../utils/constants';
-import { seafileAPI } from '../../utils/seafile-api';
-import { Utils } from '../../utils/utils';
-import Icon from '../icon';
+import ConfirmUnlinkDeviceDialog from '@/components/dialog/confirm-unlink-device';
+import Icon from '@/components/icon';
+import MobileItemMenu from '@/components/mobile-item-menu';
+import toaster from '../../../components/toast';
+import { gettext } from '../../../utils/constants';
+import { seafileAPI } from '../../../utils/seafile-api';
+import { Utils } from '../../../utils/utils';
 
 dayjs.extend(relativeTime);
