+++ frontend/src/pages/settings/social-login-saml/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Modal, ModalBody, ModalFooter, Button } from 'reactstrap';
+import ModalPortal from '@/components/modal-portal';
 import SeahubModalHeader from '@/components/seahub-modal-header';
-import { gettext, siteRoot } from '../../utils/constants';
-import ModalPortal from '../modal-portal';
+import { gettext, siteRoot } from '../../../utils/constants';
 
 const {
   csrfToken,
