+++ frontend/src/pages/settings/social-login/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import { gettext, siteRoot } from '../../utils/constants';
-import ConfirmDisconnectWechat from '../dialog/confirm-disconnect-wechat';
-import ModalPortal from '../modal-portal';
+import ConfirmDisconnectWechat from '@/components/dialog/confirm-disconnect-wechat';
+import ModalPortal from '@/components/modal-portal';
+import { gettext, siteRoot } from '../../../utils/constants';
 
 const {
   csrfToken,
