+++ frontend/src/pages/settings/web-api-auth-token/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import { gettext } from '../../utils/constants';
-import { seafileAPI } from '../../utils/seafile-api';
-import { Utils } from '../../utils/utils';
-import Icon from '../icon';
-import toaster from '../toast';
+import Icon from '@/components/icon';
+import toaster from '@/components/toast';
+import { gettext } from '../../../utils/constants';
+import { seafileAPI } from '../../../utils/seafile-api';
+import { Utils } from '../../../utils/utils';
 
 class WebAPIAuthToken extends React.Component {
